+++ templates/expo-template-default/src/components/app-tabs.tsx
